refactor(control_performance_analysis): rework parameters (#4730)
* refactor the configuration files of the node control_performance_analysis according to the new ROS node config guideline.
Rename controller_performance_analysis.launch.xml->control_performance_analysis.launch.xml
update the parameter information in the README.md


* style(pre-commit): autofix

* Update the type of a parameter in schema file.


* revert copyright info


---------

+++ control/control_performance_analysis/src/control_performance_analysis_node.cpp
@@ -47,7 +47,7 @@ ControlPerformanceAnalysisNode::ControlPerformanceAnalysisNode(
   // Node Parameters.
   param_.curvature_interval_length_ = declare_parameter<double>("curvature_interval_length");
   param_.prevent_zero_division_value_ = declare_parameter<double>("prevent_zero_division_value");
-  param_.odom_interval_ = declare_parameter<int>("odom_interval");
+  param_.odom_interval_ = declare_parameter<int64_t>("odom_interval");
   param_.acceptable_max_distance_to_waypoint_ =
     declare_parameter<double>("acceptable_max_distance_to_waypoint");
   param_.acceptable_max_yaw_difference_rad_ =
